requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
Django==3.2.12
django-cors-headers==3.10.1
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.4.0
Faker==4.0.0
idna==3.3
importlib-metadata==4.8.3
Markdown==3.3.6
mixer==7.0.6
pycparser==2.21
PyJWT==2.3.0
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
typing_extensions==4.1.1
urllib3==1.26.9
zipp==3.6.0
